+++ app/models/surveys/question.rb
@@ -1,5 +1,22 @@
 class Question < ActiveRecord::Base
+  # Constants
+  DISPLAY_TYPE = ["custom_date_input", "radio_input", "checkbox_input", "height_input", "number_input", "radio_input_multiple"]
+
+  # Concerns
+  include Localizable
+  include Votable
+  include Deletable
+
+  # Translations
+  localize :text
+
+  # Model Validation
+  validates_presence_of :text_en, :slug, :user_id, :display_type
+  validates_uniqueness_of :slug, scope: [ :deleted ]
+  validates_format_of :slug, with: /\A(?!\Anew\Z)[a-z][a-z0-9\-]*\Z/
+
   # Associations
+  belongs_to :user
   has_and_belongs_to_many :answer_templates, -> { order "answer_templates.created_at asc" }
   belongs_to :group
   has_many :answers, -> { where deleted: false }
@@ -8,14 +25,18 @@ class Question < ActiveRecord::Base
   has_many :votes
   has_many :reports
 
-  # Concerns
-  include Localizable
-  include Votable
-
-  localize :text
-
   ## DAG
   has_dag_links :link_class_name => 'QuestionEdge'
+
+  # Class Methods
+
+  def to_param
+    slug.blank? ? id : slug
+  end
+
+  def self.find_by_param(input)
+    self.where("questions.slug = ? or questions.id = ?", input.to_s, input.to_i).first
+  end
 
   def next_question(survey)
     candidate_edges(survey).first