+++ db/migrate/20150731175910_add_deleted_to_questions.rb
@@ -1,0 +1,5 @@
+class AddDeletedToQuestions < ActiveRecord::Migration
+  def change
+    add_column :questions, :deleted, :boolean, null: false, default: false
+  end
+end